% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_species_download_sub_function.R
\name{snake_antivenoms_get}
\alias{snake_antivenoms_get}
\title{Download Antivenoms For Snake Species}
\usage{
snake_antivenoms_get(snake_species)
}
\arguments{
\item{snake_species}{Used by the function to work out which snake_id it needs to look at}
}
\description{
This sub-function is used by the function snake_dowload. To clean stuff up from the rvest output. This one is quite a bit different as it creates a data_frame, with one line for each antivenom
}
\examples{
snake_previous_name_get()
}
