% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_country_download_function.R
\name{snake_country_download}
\alias{snake_country_download}
\title{This function is used by the snake_country_data script, to create a data_frame listing all countries and the snakes they contain}
\usage{
snake_country_download(x, country_id, country_name)
}
\arguments{
\item{x}{the address for selenium to find the next country}

\item{country_id}{the ID WHO uses for countries}

\item{country_name}{the name of the country}
}
\description{
This function is used by the snake_country_data script, to create a data_frame listing all countries and the snakes they contain
}
\examples{
snake_country_download()
}
