% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_country_download_sub_function.R
\name{snake_country_secondary_download}
\alias{snake_country_secondary_download}
\title{This function is used by the snake_country_download script, to create a data_frame listing the snakes on an individual page}
\usage{
snake_country_secondary_download(page_element)
}
\arguments{
\item{page_element}{the next page to load to download the snake country data}
}
\description{
This function is used by the snake_country_download script, to create a data_frame listing the snakes on an individual page
}
\examples{
snake_country_secondary_download()
}
