% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_species_download_function.R
\name{snake_download}
\alias{snake_download}
\title{Download Individual Snake Species}
\usage{
snake_download(snake_species_id)
}
\arguments{
\item{snake_species_id}{Used by the script to work out which snake_id it needs to look at}
}
\description{
This function is used by the snake_species_data_frame script, to create a data_frame with a single line about a single snake
}
\examples{
snake_download()
}
