% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_species_download_sub_function.R
\name{snake_family_get}
\alias{snake_family_get}
\title{Download Family Name Of Snake Species}
\usage{
snake_family_get(snake_species)
}
\arguments{
\item{snake_species}{Used by the function to work out which snake_id it needs to look at}
}
\description{
This sub-function is used by the function snake_dowload. To clean stuff up from the rvest output.
}
\examples{
snake_family_get()
}
