% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{snake_species_data}
\alias{snake_species_data}
\title{237 Medically Relevant Snakes}
\format{A data frame with 237 rows and 12 variables, one variable is a list, two variables are nested:
\describe{
  \item{snake_species_id}{WHO unique identifier for snake}
  \item{snake_common_name}{Name of the Snake Species, e.g. "Common Death Adder"}
  \item{snake_species}{Scientific Name of the Snake Species, e.g. "Acanthophis antarcticus"}
  \item{snake_family}{Name of the Species Family, e.g. "Elapidae"}
  \item{snake_map_link}{URL of WHO Map of the Snake Geographical Distribution}
  \item{snake_map_legend_link}{URL for the legend for the snake_map_link}
  \item{snake_picture_link}{URL of WHO Picture of the Snake}
  \item{snake_common_name_other}{Other Name of the Snake Species}
  \item{snake_common_name_remainder}{List of Other Names of the Snake Species}
  \item{snake_previous_name}{Any Previous Scientific Name of the Snake Species}
  \item{snake_regions}{Nested Data Frame of all Regions and Sub-Regions}
  \item{snake_antivenoms}{Nested Data Frame of Antivenoms, Manufacturers, Countries of Origin}
  
  ...
}}
\source{
\url{http://apps.who.int/bloodproducts/snakeantivenoms/database/}
}
\usage{
snake_species_data
}
\description{
A dataset containing the names, regions, and antivenoms for the 237 medically relevant snake species on the WHO online database
}
\keyword{datasets}
